updated layout lesson with caption image files
updated layout lesson with caption image raw file dump. this lesson will show me how to position a caption over an image using relative and absolute positioning. If I had to guess, i may be placing one item on top of the other using something similar to a float, and then using relative position for the parent, and absolute position for the caption itself.