uploaded input element raw files
uploaded input elements raw files and also updated notes for form element lesson. No forms nested within forms. Two attributes needed for form tag. First is action which in this case is "index.html", and the second is method, which in this case is "post" or it can also be "get"